code structure modification.
