+++ server/index.js
@@ -8,7 +8,7 @@ var appCommons = require('onf-core-model-ap/applicationPattern/commons/AppCommon
 var serverPort = 8080;
 
 // uncomment if you do not want to validate security e.g. operation-key, basic auth, etc
-//appCommons.openApiValidatorOptions.validateSecurity = false;
+appCommons.openApiValidatorOptions.validateSecurity = false;
 
 // swaggerRouter configuration
 var options = {
